be: Host endpoints, Store SSH sessions in HashMap

+++ backend/src/main/java/dev/wybran/perceptus/repository/HostRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HostRepository extends JpaRepository<Host, Integer> {
+public interface HostRepository extends JpaRepository<Host, Long> {
 }
